requirements.txt
anomalo
google-cloud-bigquery
google-cloud-dataplex
python-dotenv
requests
